CodeGen: Create independent basic blocks.
Before code generating the clast, create independent basic blocks. This means
that every scalar operand that is used in a block and that has a value which can
be described using SCEV is calculated in the block where it is used.
The value is calculated by expanding by expanding the its SCEV.

Independent basic blocks have the advantage to only depend on loop ivs and
parameters. Therefore they can be scheduled in arbitrary order.