Dependences: Detect parallel loops based on distance vectors
This detection is order of magnitudes faster. The implementation is yet
incomplete as we currently only track flow dependences. For those such
an approach is not sufficient. Therefore, currently it may happen that we
assume a loop is parallel, but is it actually not. I believe those cases
will probably not appear to often.