Pocc/pollycc: Allow more vectorization when -fvector is enabled.
We now detect the innermost loop and strip-mine it by four. This will yield in
many cases to innermost loops with just four iterations, which we can than
vectorize. In case those loop are not parallel or contain constructs like
for (j = max(i,x); j < (i+3); j++) the number of loop iterations is not
statically known and we are are not able to create vector code. In this cases
the non optimal innermost loops remain.